profileimage and upload document updated

+++ RentalTracker.Web/DataAccessLayer/CommonService.cs
@@ -3,6 +3,8 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using RentalTracker.Web.Areas.Identity.Data;
 using RentalTracker.Web.Models;
+using SixLabors.ImageSharp;
+using SixLabors.ImageSharp.Processing;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -23,7 +25,7 @@ namespace RentalTracker.Web.DAL
             _webHostEnvironment = webHostEnvironment;
         }
 
-        public async Task<City> CityById(int? id)
+        public async Task<City> CityByIdAsync(int? id)
         {
             if (id == null)
                 return null;
@@ -31,7 +33,7 @@ namespace RentalTracker.Web.DAL
             return await _repository.Cities.FindAsync(id);
         }
 
-        public async Task<State> StateById(int? id)
+        public async Task<State> StateByIdAsync(int? id)
         {
             if (id == null)
                 return null;
@@ -39,7 +41,7 @@ namespace RentalTracker.Web.DAL
             return await _repository.States.FindAsync(id);
         }
 
-        public async Task<Country> CountryById(int? id)
+        public async Task<Country> CountryByIdAsync(int? id)
         {
             if (id == null)
                 return null;
@@ -87,13 +89,29 @@ namespace RentalTracker.Web.DAL
             return await _repository.Cities.ToListAsync();
         }
 
-        public async Task<ICollection<City>> CityLists(int? stateId)
+        public async Task<ICollection<City>> CityListByStateIdAsync(int? stateId)
         {
             if (stateId == null)
                 return await _repository.Cities.ToListAsync();
             else
                 return await _repository.Cities.Where(p => p.StateId == stateId).ToListAsync();
         }
+
+        public async Task<ICollection<City>> GetCityStateCountryLists(int? cityId)
+        {
+            var cities = await _repository.Cities
+                .Include(p => p.State)
+                .Where(c => c.CityId == cityId)
+                .ToListAsync();
+
+            //if (stateId == null)
+            //    return await _repository.Cities.ToListAsync();
+            //else
+            //    return await _repository.Cities.Where(p => p.StateId == stateId).ToListAsync();
+
+            return cities; //await _repository.Cities.ToListAsync();
+        }
+
 
         public void DeleteFileIfExists(string fileUrl, Directories directories)
         {
@@ -111,21 +129,70 @@ namespace RentalTracker.Web.DAL
 
         public string UploadedFile(string Url, IFormFile formFile, Directories directoryName)
         {
-            if (Url != null)
-            {
-                var path = Url == null ? null :
-                    Path.Combine(_webHostEnvironment.WebRootPath, directoryName.ToString() + "/", Url);
-
-                if (File.Exists(path))
-                {
-                    File.Delete(path);
-                }
-            }
-
             string uniqueFileName = null;
 
             if (formFile != null)
             {
+                if (Url != null)
+                {
+                    var path = Url == null ? null :
+                        Path.Combine(_webHostEnvironment.WebRootPath, directoryName.ToString() + "/", Url);
+
+                    if (File.Exists(path))
+                    {
+                        File.Delete(path);
+                    }
+                }
+
+                string uploadsFolder = Path.Combine(_webHostEnvironment.WebRootPath, directoryName.ToString() + "/");
+
+                if (!Directory.Exists(uploadsFolder))
+                    Directory.CreateDirectory(uploadsFolder);
+
+                uniqueFileName = Guid.NewGuid().ToString() + "_" + formFile.FileName;
+                
+                var extension = Path.GetExtension(formFile.FileName);
+                string filePath = Path.Combine(uploadsFolder, uniqueFileName);
+                
+                if (extension.ToLower() == ".xlsx" || extension.ToLower() == ".pdf")
+                {
+                    using (var fileStream = new FileStream(filePath, FileMode.Create))
+                    {
+                        formFile.CopyTo(fileStream);
+                    }
+                }
+                else
+                {
+                    using (var image = Image.Load(formFile.OpenReadStream()))
+                    {
+                        int[] size = ResizeImage(image, 200, 200);
+                        image.Mutate(p => p.Resize(size[0], size[1]));
+                        image.Save(filePath);
+                    }
+                }
+
+            }
+
+            return uniqueFileName;
+        }
+
+        public string UploadedFile(string Url, IFormFile formFile, Directories directoryName, int width, int height)
+        {
+            string uniqueFileName = null;
+
+            if (formFile != null)
+            {
+                if (Url != null)
+                {
+                    var path = Url == null ? null :
+                        Path.Combine(_webHostEnvironment.WebRootPath, directoryName.ToString() + "/", Url);
+
+                    if (File.Exists(path))
+                    {
+                        File.Delete(path);
+                    }
+                }
+
                 string uploadsFolder = Path.Combine(_webHostEnvironment.WebRootPath, directoryName.ToString() + "/");
 
                 if (!Directory.Exists(uploadsFolder))
@@ -135,13 +202,66 @@ namespace RentalTracker.Web.DAL
 
                 string filePath = Path.Combine(uploadsFolder, uniqueFileName);
 
-                using (var fileStream = new FileStream(filePath, FileMode.Create))
+                var extension = Path.GetExtension(formFile.FileName);
+
+                if (extension.ToLower() == ".xlsx" || extension.ToLower() == ".pdf")
                 {
-                    formFile.CopyTo(fileStream);
+                    using (var fileStream = new FileStream(filePath, FileMode.Create))
+                    {
+                        formFile.CopyTo(fileStream);
+                    }
                 }
+                else
+                {
+                    using (var image = Image.Load(formFile.OpenReadStream()))
+                    {
+                        int[] size = ResizeImage(image, width, height);
+                        image.Mutate(p => p.Resize(size[0], size[1]));
+                        image.Save(filePath);
+                    }
+                }
+
             }
 
             return uniqueFileName;
+        }
+
+        public int[] ResizeImage(Image img, int maxWidth, int maxHeight)
+        {
+            int[] val = new int[2];
+
+            if (img.Width > maxWidth || img.Height > maxHeight)
+            {
+                double widthRatio = (double)img.Width / (double)maxWidth;
+                double heightRatio = (double)img.Height / (double)maxHeight;
+
+                double ratio = Math.Max(widthRatio, heightRatio);
+                int newWidth = (int)(img.Width / ratio);
+                int newHeight = (int)(img.Height / ratio);
+
+                val[0] = newWidth;
+                val[1] = newHeight;
+
+                return val;
+                //return newHeight.ToString() + "," + newWidth.ToString();
+            }
+            else
+            {
+                val[0] = img.Width;
+                val[1] = img.Height;
+
+                return val;
+
+                //return img.Height.ToString() + "," + img.Width.ToString();
+            }
+        }
+
+        public async Task<ICollection<State>> StateListByCountryIdAsync(int? countryId)
+        {
+            if (countryId == null)
+                return await _repository.States.ToListAsync();
+            else
+                return await _repository.States.Where(p => p.CountryId == countryId).ToListAsync();
         }
     }
 }
